refactor: update types
