+++ app/constants/pageBlocks.ts
@@ -6,5 +6,7 @@ export const BLOCK_IDS = {
   INTRO_SECTION: 'IntroSection',
   LIATOSHYNSKY_FOUNDATION: 'FoundationInfo',
   LIATOSHYNSKY_OFFICE: 'LiatoshynskyOffice',
-  OUR_MISSION: 'OurMission'
+  OUR_MISSION: 'OurMission',
+  OUR_GOALS: 'OurGoals',
+  WHAT_WE_DO: 'WhatWeDo'
 } as const;
