+++ app/types/store/pages/about-us/blocks/liatoshynskyFoundationBlock.ts
@@ -1,32 +1,4 @@
-export type ProseTextNode = {
-  type: 'text';
-  text: string;
-  marks?: { type: 'bold' | 'italic' }[];
-};
-
-export type ParagraphNode = {
-  type: 'paragraph';
-  content: ProseTextNode[];
-};
-
-export type HeadingNode = {
-  type: 'heading';
-  level: 1 | 2 | 3;
-  content: ProseTextNode[];
-};
-
-export type ImageNode = {
-  type: 'image';
-  src: string;
-  alt?: string;
-};
-
-export type ProseNode = ParagraphNode | HeadingNode | ImageNode;
-
-export type ProseDoc = {
-  type: 'doc';
-  content: ProseNode[];
-};
+import { ProseDoc } from '~/types/common';
 
 export type ImageType = {
   src: string;
